Create bed6_to_bed12.sh
Using a bed file of genomic coordinates, this script will use  `bedtools intersect` to find the overlapping features in the region, producing a `.gtf` file. If the region is intronic, this `.gtf` file will be empty and the bed12 file for the intron will be created manually. If the `.gtf` file has features (i.e exons) then it is converted to bed12 format using UCSC tools. A strict overlap using the start and stop sites of the genomic region ensure that the correct exons are retained. 

The only "dodgy" part is that I simply take the top transcript and use it as the representative sequence. I think the top transcript (head -n 1) has all exons present in it...